Shop UI 수정#3
DB 연동 확인 중간단계
-UIFRShop 확인

+++ Assets/Scripts/ScriptsKDH/UIListItem.cs
@@ -12,7 +12,7 @@ public class UIListItem : MonoBehaviour
     public TMP_Text price;
     public Button btn;
 
-    public void Init(Sprite sp, string itemName, string itemDescription, int price)
+    public void Init(Sprite sp, string itemName, string itemDescription, long price)
     {
         icon.sprite = sp;
         this.itemName.text = itemName;
